Add GitHub Actions workflow for running tests and update okteto.yml
This commit adds a new GitHub Actions workflow file, tests.yaml, which is triggered on pull requests to the main branch. The workflow sets up the development environment, checks out the code, sets environment variables, and runs tests using the okteto/test action. Additionally, the okteto.yml file is updated to include a new test configuration for the worker service, using the golang image and running a command to print environment variables related to GitHub.

Signed-off-by: Javier Lopez <javier@okteto.com>